requirements.txt
streamlit
tensorflow_text
Pillow>=6.2.2
tensorflow-cpu
numpy
pytest-shutil
tensorflow_hub
beautifulsoup4

